fixed #20 and minor changes
