+++ config.ini
@@ -1,6 +1,6 @@
 [infura]
 INFURA_NODE_LINK = https://mainnet.infura.io/v3/
-INFURA_TEST_NODE_LINK = https://goerli.infura.io/v3/
+INFURA_TEST_NODE_LINK = https://sepolia.infura.io/v3/
 INFURA_PID = 
 INFURA_Test_PID = 
 RPC_POLY_TEST_NODE_LINK = https://polygon-mumbai.g.alchemy.com/v2/
@@ -18,9 +18,9 @@ ETHERSCAN_API_KEY =
 TRANSACTION_LINK = https://api.etherscan.io/api?module=account&action=txlist&address={}&startblock=0&endblock=99999999&page={}&offset=100&sort=asc&apikey={}
 INTERNAL_TRANSACTION_LINK = https://api.etherscan.io/api?module=account&action=txlistinternal&address={}&m=advanced&startblock=0&endblock=99999999&page={}&offset=100&sort=asc&apikey={}
 EVENT_LINK = https://api.etherscan.io/api?module=logs&action=getLogs&address={}&fromBlock=0&toBlock=99999999&page={}&offset=100&apikey={}
-Test_TRANSACTION_LINK = https://api-goerli.etherscan.io/api?module=account&action=txlist&address={}&startblock=0&endblock=99999999&page={}&offset=100&sort=asc&apikey={}
-Test_INTERNAL_TRANSACTION_LINK = https://api-goerli.etherscan.io/api?module=account&action=txlistinternal&address={}&m=advanced&startblock=0&endblock=99999999&page={}&offset=100&sort=asc&apikey={}
-TEST_EVENT_LINK = https://api-goerli.etherscan.io/api?module=logs&action=getLogs&address={}&fromBlock=0&toBlock=99999999&page={}&offset=100&apikey={}
+Test_TRANSACTION_LINK = https://api-sepolia.etherscan.io/api?module=account&action=txlist&address={}&startblock=0&endblock=99999999&page={}&offset=100&sort=asc&apikey={}
+Test_INTERNAL_TRANSACTION_LINK = https://api-sepolia.etherscan.io/api/api?module=account&action=txlistinternal&address={}&m=advanced&startblock=0&endblock=99999999&page={}&offset=100&sort=asc&apikey={}
+TEST_EVENT_LINK = https://api-sepolia.etherscan.io/api?module=logs&action=getLogs&address={}&fromBlock=0&toBlock=99999999&page={}&offset=100&apikey={}
 
 [polygonscan]
 POLYGONSCAN_API_KEY =
