+++ src/key_approx_analysis/key_approx_analyzer.py
@@ -4,7 +4,7 @@ import pprint
 import itertools
 from slither.slither import Slither
 from configparser import ConfigParser
-from src.ast_parsing.ast_parser import generate_ast
+from src.ast_parsing.ast_parser import generate_ast, get_contract_details, get_contract_details_new
 from src.ast_parsing.ast_parser import parse_ast
 from web3.auto import Web3
 from web3.middleware import geth_poa_middleware
@@ -147,488 +147,6 @@ def get_vars(expr):
     else:
         return_vars.append(vars_split[0].strip())
     return return_vars
-
-def find_diamond_for_class(tree, class_name):
-    # A recursive function to get all base classes for a given class
-    def all_bases(cls, accum=None):
-        if accum is None:
-            accum = {}
-        bases = tree.get(cls, [])
-        for base in bases:
-            accum[base] = accum.get(base, 0) + 1
-            all_bases(base, accum)
-        return accum
-
-    # Find if the specified class has any base classes included multiple times
-    bases = all_bases(class_name)
-    diamond_bases = {base: count for base, count in bases.items() if count > 1}
-    return diamond_bases
-
-
-def merge(sequences):
-    """
-    Merges multiple sequences into a single C3 linearization, as per the C3 algorithm.
-    """
-    result = []
-    while True:
-        non_empty_seqs = [seq for seq in sequences if seq]
-        if not non_empty_seqs:
-            return result
-        for seq in non_empty_seqs:  # find merge candidates among seq heads
-            candidate = seq[0]
-            if not any(candidate in s[1:] for s in non_empty_seqs):
-                break
-        else:
-            raise Exception("Inconsistent hierarchy")
-        result.append(candidate)
-        for seq in non_empty_seqs:  # remove candidate
-            if seq[0] == candidate:
-                del seq[0]
-
-def c3_linearization(classname, tree):
-    """
-    Calculates the C3 linearization for a given class.
-    """
-    base_classes = tree.get(classname, [])
-    sequences = [c3_linearization(base_class, tree) for base_class in base_classes] + [base_classes + [classname]]
-    return merge(sequences)
-
-def unroll_struct(struct, all_contract_dict):
-    """Takes in a struct data type, returns list of variables in the struct"""
-    var_lst = []
-    for var_struct in struct['members']:
-        var_lst.append(format_variable(var_struct, all_contract_dict))
-    return var_lst
-
-def format_variable(var_struct, all_contracts_dict):
-    """Takes in a variable struct, returns the formatted variable according to it's type"""
-    if var_struct['typeName']['type'] == 'ElementaryTypeName':
-        var_dict = {}
-        var_dict['type'] = var_struct['typeName']['type']
-        var_dict['dataType'] = var_struct['typeName']['name']
-        var_dict['name'] = var_struct['name']
-        return var_dict
-    elif var_struct['typeName']['type'] == 'UserDefinedTypeName':      
-        # if userdefined variable is an enum, treat it as elementary type
-        if "." in var_struct['typeName']['namePath']:
-            var_struct['typeName']['namePath'] = var_struct['typeName']['namePath'].split(".")[-1]
-        try:
-            var_struct_data_type = all_contracts_dict[var_struct['typeName']['namePath']]['vars'][0]['dataType']
-        except:
-            var_struct_data_type = ''
-        if var_struct_data_type == 'enum':
-            var_dict = {}
-            var_dict['type'] = 'ElementaryTypeName'
-            var_dict['dataType'] = 'enum'
-            var_dict['name'] = var_struct['name']
-        else:
-            var_dict = {}
-            var_dict['type'] = var_struct['typeName']['type']
-            var_dict['dataType'] = var_struct['typeName']['namePath']
-            try:
-                var_dict['typeVars'] = all_contracts_dict[var_struct['typeName']
-                    ['namePath']]['vars']
-            except:
-                var_dict['typeVars'] = []
-        var_dict['name'] = var_struct['name']
-        return var_dict
-    elif var_struct['typeName']['type'] == 'Mapping':
-        var_dict = {}
-        var_dict['type'] = var_struct['typeName']['type']
-        var_dict['keyType'] = var_struct['typeName']['keyType']
-        var_dict['valueType'] = var_struct['typeName']['valueType']
-        var_dict['name'] = var_struct['name']
-        return var_dict
-    elif var_struct['typeName']['type'] == 'ArrayTypeName':
-        name_struct = var_struct['typeName']['baseTypeName']
-        lens = []
-        try:
-            lens.append(var_struct['typeName']['length']['number'])
-        except:
-            try:
-                lens.append(var_struct['typeName']['length']['name'])
-            except:
-                lens.append(var_struct['typeName']['length'])
-        # iterate over all dimensions of array to get length of each dimension
-        while not ('name' in name_struct or 'namePath' in name_struct):
-            if 'length' in name_struct:
-                try:
-                    lens.append(name_struct['length']['number'])
-                except:
-                    try:
-                        lens.append(name_struct['length']['name'])
-                    except:
-                        lens.append(name_struct['length'])
-            name_struct = name_struct['baseTypeName']
-        new_lens = []
-        for i in range(1, len(lens)+1):
-            new_lens.append(lens[-i])
-        type_struct = name_struct['type']
-        var_dict = {}
-        if 'name' in name_struct:
-            name_struct = name_struct['name']
-            type_struct = 'ElementaryTypeName'
-        else:
-            name_struct = name_struct['namePath']
-            type_struct = 'UserDefinedTypeName'
-
-        var_dict['type'] = var_struct['typeName']['type']
-        var_dict['dataTypeType'] = type_struct
-        var_dict['dataTypeName'] = name_struct
-        var_dict['length'] = new_lens
-        var_dict['name'] = var_struct['name']
-        var_dict['curr'] = -1
-        if len(lens) > 1:
-            var_dict['dimension'] = 'multi'
-        else:
-            var_dict['dimension'] = 'single'
-        if None not in lens:
-            var_dict['StorageType'] = 'static'
-        else:
-            var_dict['StorageType'] = 'dynamic'
-        return var_dict
-
-
-def variable_unrolling(subnodes, all_contracts_dict, all_vars):
-    statevars = []
-    for node in subnodes:
-        if node['type'] == 'StateVariableDeclaration':
-            vars = node['variables']
-            for variable in vars:
-                # grab the value if the value is assigned to a variable at compile time
-                if 'isDeclaredConst' in variable:
-                    if variable['isDeclaredConst'] == True:
-                        try:
-                            if 'value' in variable['expression']:
-                                all_vars.append(
-                                    [variable['name'], variable['typeName']['name'], variable['expression']['value']])
-                            elif 'number' in variable['expression']:
-                                all_vars.append(
-                                    [variable['name'], variable['typeName']['name'], variable['expression']['number']])
-                        except:
-                            pass
-                        continue
-                if 'isDeclaredImmutable' in variable:
-                    if variable['isDeclaredImmutable'] == True:
-                        try:
-                            if 'value' in variable['expression']:
-                                all_vars.append(
-                                    [variable['name'], variable['typeName']['name'], variable['expression']['value']])
-                            elif 'number' in variable['expression']:
-                                all_vars.append(
-                                    [variable['name'], variable['typeName']['name'], variable['expression']['number']])
-                        except:
-                            pass
-                        continue
-                st_var = statevars.append(format_variable(variable, all_contracts_dict))
-                if st_var != None:
-                    statevars.append(st_var)
-        elif node['type'] == 'StructDefinition':
-            tmp = unroll_struct(node, all_contracts_dict)
-            all_contracts_dict[node['name']] = {'vars': tmp}
-        elif node['type'] == 'EnumDefinition':
-            var_dict = {}
-            var_dict['type'] = 'ElementaryTypeName'
-            var_dict['dataType'] = 'enum'
-            var_dict['name'] = node['name']
-            all_contracts_dict[node['name']] = {'vars': [var_dict]}
-    return statevars, all_contracts_dict, all_vars
-
-
-def get_contract_details(children, contract_name):
-    """
-    Takes AST of the source code, and returns details of all defined contracts, structs and enums.
-
-    Parameters:
-        contracts (object): AST of source code containing AST of all contracts.
-
-    Returns:
-        all_vars (list): list of all state variables.
-        all_contracts_dict (dict): contains details of all contracts defined in the source code.
-    """
-    all_vars = []
-    all_contracts_dict = {}
-    inherit_tree = {}
-    linearized_inherit_tree = {}
-    for contract in children:
-        if contract == None:
-            continue
-        try:
-            sub_nodes = contract['subNodes']
-        except:
-            sub_nodes = []
-        if contract['type'] == 'PragmaDirective':
-            continue
-        if 'baseContracts' in contract:
-            if contract['baseContracts'] != []:
-                lst = contract['baseContracts']
-                parent_list = []
-                for basecontract in lst:
-                    parent_list.append(basecontract['baseName']['namePath'])
-                inherit_tree[contract['name']] = parent_list
-                
-    diamonds = find_diamond_for_class(inherit_tree, contract_name)
-    if diamonds:
-        print("************************************")
-        print(f"Diamond inheritance detected for class {contract_name}: {diamonds}")
-        print("************************************")
-    else:
-        print(f"No diamond inheritance detected for class {contract_name}.")
-    for cont in inherit_tree:
-        temp = c3_linearization(cont, inherit_tree)
-        linearized_inherit_tree[cont] = temp[:-1]
-    # print(linearized_inherit_tree)
-
-    for contract in children:
-        parent = []
-        if contract == None:
-            continue
-        try:
-            sub_nodes = contract['subNodes']
-        except:
-            sub_nodes = []
-        if contract['type'] == 'PragmaDirective':
-            continue
-        state_vars, all_contracts_dict, all_vars = variable_unrolling(
-            sub_nodes, all_contracts_dict, all_vars)
-        if 'baseContracts' in contract:
-            if contract['baseContracts'] != []:
-                    lst = linearized_inherit_tree[contract['name']]
-                    parent_vars = []
-                    for parent_cont_name in lst:
-                        if len(all_contracts_dict) != 0:
-                            b_statevars = all_contracts_dict[parent_cont_name]['vars']
-                            for var in b_statevars:
-                                if type(var) == list:
-                                    for va in var:
-                                        if va not in parent_vars:
-                                            parent_vars.append(va)
-                                else:
-                                    if var not in parent_vars:
-                                        parent_vars.append(var)
-                    state_vars = parent_vars + state_vars
-                    parent = lst
-        try:
-            all_contracts_dict[contract['name']] = {
-                'vars': state_vars, 'parent': parent, 'type': contract['type']}
-        except:
-            continue
-    return all_vars, all_contracts_dict, diamonds
-
-def format_variable_new(var_struct, all_contracts_dict):
-    """Takes in a variable struct, returns the formatted variable according to it's type"""
-    
-    if var_struct['typeName']['nodeType'] == 'ElementaryTypeName':
-        var_dict = {}
-        var_dict['type'] = var_struct['typeName']['nodeType']
-        var_dict['dataType'] = var_struct['typeName']['name']
-        var_dict['name'] = var_struct['name']
-        return var_dict
-    elif var_struct['typeName']['nodeType'] == 'UserDefinedTypeName':
-        if 'enum' in var_struct['typeDescriptions']['typeIdentifier']:
-            var_dict = {}
-            var_dict['type'] = 'ElementaryTypeName'
-            var_dict['dataType'] = 'enum'
-            var_dict['name'] = var_struct['name']
-        else:
-            var_dict = {}
-            var_dict['type'] = var_struct['typeName']['nodeType']
-            try:
-                var_dict['dataType'] = var_struct['typeName']['pathNode']['name']
-            except:
-                var_dict['dataType'] = var_struct['typeName']['name']
-            # var_dict['defType'] = all_contracts_dict[var_dict['dataType']]['type']
-            var_dict['name'] = var_struct['name']
-            try:
-                var_dict['typeVars'] = all_contracts_dict[var_dict['dataType']]['vars']
-            except:
-                var_dict['typeVars'] = []
-        return var_dict
-    elif var_struct['typeName']['nodeType'] == 'Mapping':
-        var_dict = {}
-        var_dict['type'] = var_struct['typeName']['nodeType']
-        var_dict['keyType'] = var_struct['typeName']['keyType']
-        var_dict['valueType'] = var_struct['typeName']['valueType']
-        var_dict['name'] = var_struct['name']
-        return var_dict
-    elif var_struct['typeName']['nodeType'] == 'ArrayTypeName':
-        name_struct = var_struct['typeName']['baseType']
-        lens = []
-        try:
-            if var_struct['typeName']['length']['nodeType'] == 'Literal':
-                lens.append(var_struct['typeName']['length']['value'])
-            else:
-                lens.append(var_struct['typeName']['length']['name'])
-        except:
-            try:
-                lens.append(var_struct['typeName']['length'])
-            except:
-                lens.append(None)
-        # iterate over all dimensions of array to get length of each dimension
-        while not ('name' in name_struct or 'namePath' in name_struct):
-            try:
-                if name_struct['length']['nodeType'] == 'Literal':
-                    lens.append(name_struct['length']['value'])
-                else:
-                    lens.append(name_struct['length']['name'])
-            except:
-                try:
-                    lens.append(name_struct['length'])
-                except:
-                    lens.append(None)
-            name_struct = name_struct['baseType']
-        type_struct = name_struct['nodeType']
-        
-        new_lens = []
-        for x in range(1, len(lens)+1):
-            new_lens.append(lens[-x])
-        
-        var_dict = {}
-        var_dict['type'] = var_struct['typeName']['nodeType']
-        var_dict['dataTypeType'] = type_struct
-        var_dict['dataTypeName'] = name_struct['name']
-        var_dict['length'] = new_lens
-        var_dict['name'] = var_struct['name']
-        var_dict['curr'] = -1
-        if len(lens) > 1:
-            var_dict['dimension'] = 'multi'
-        else:
-            var_dict['dimension'] = 'single'
-        if None not in lens:
-            var_dict['StorageType'] = 'static'
-        else:
-            var_dict['StorageType'] = 'dynamic'
-        return var_dict
-
-def unroll_struct_new(struct, all_contract_dict):
-    """Takes in a struct data type, returns list of variables in the struct"""
-    var_lst = []
-    for var_struct in struct['members']:
-        var_lst.append(format_variable_new(var_struct, all_contract_dict))
-    return var_lst
-
-
-def variable_unrolling_new(subnodes, all_contracts_dict, all_vars):
-    statevars = []
-    for node in subnodes:
-        if node['nodeType'] == "VariableDeclaration":
-          if node['stateVariable'] == True:
-            try:
-                if node['constant'] == True or node['mutability'] == "immutable":
-                    # grab the value if the value is assigned to a variable at compile time
-                    try:
-                        if 'number' in node['value']['kind'] or 'string' in node['value']['kind']:
-                            all_vars.append(
-                                [node['name'], node['typeDescriptions']['typeString'], node['value']['value']])
-                            continue
-                    except:
-                        continue
-                else:
-                    st_var = format_variable_new(node, all_contracts_dict)
-                    if st_var != None:
-                        statevars.append(st_var)
-            except:
-                if node['constant'] == True:
-                    # grab the value if the value is assigned to a variable at compile time
-                    try:
-                        if 'number' in node['value']['kind'] or 'string' in node['value']['kind']:
-                            all_vars.append(
-                                [node['name'], node['typeDescriptions']['typeString'], node['value']['value']])
-                            continue
-                    except:
-                        continue
-                else:
-                    st_var = format_variable_new(node, all_contracts_dict)
-                    if st_var != None:
-                        statevars.append(st_var)
-        elif node['nodeType'] == 'StructDefinition':
-            tmp = unroll_struct_new(node, all_contracts_dict)
-            all_contracts_dict[node['name']] = {'vars': tmp, 'type': node['nodeType']}
-        elif node['nodeType'] == 'EnumDefinition':
-            var_dict = {}
-            var_dict['type'] = 'ElementaryTypeName'
-            var_dict['dataType'] = 'enum'
-            var_dict['name'] = node['name']
-            all_contracts_dict[node['name']] = {'vars': [var_dict]}
-
-    return statevars, all_contracts_dict, all_vars
-
-def get_contract_details_new(contracts, contract_name):
-    """
-    Takes AST of the source code, and returns details of all defined contracts, structs and enums.
-
-    Parameters:
-        contracts (object): AST of source code containing AST of all contracts.
-
-    Returns:
-        all_vars (list): list of all state variables.
-        all_contracts_dict (dict): contains details of all contracts defined in the source code.
-    """
-    all_vars = []
-    all_contracts_dict = {}
-    inherit_tree = {}
-    linearized_inherit_tree = {}
-    for contract in contracts:
-        if contract == None:
-            continue
-        if contract['nodeType'] == 'PragmaDirective':
-            continue
-        if 'baseContracts' not in contract:
-            continue
-        if contract['baseContracts'] != []:
-            lst = contract['baseContracts']
-            parent_list = []
-            for basecontract in lst:
-                parent_list.append(basecontract['baseName']['name'])
-            inherit_tree[contract['name']] = parent_list
-    # print(inherit_tree)
-    diamonds = find_diamond_for_class(inherit_tree, contract_name)
-    if diamonds:
-        print("************************************")
-        print(f"Diamond inheritance detected for class {contract_name}: {diamonds}")
-        print("************************************")
-    else:
-        print(f"No diamond inheritance detected for class {contract_name}.")
-
-    for cont in inherit_tree:
-        temp = c3_linearization(cont, inherit_tree)
-        linearized_inherit_tree[cont] = temp[:-1]
-    # print(linearized_inherit_tree)
-
-    for contract in contracts:
-        parent = []
-        if contract == None:
-            continue
-        if contract['nodeType'] == 'PragmaDirective':
-            continue
-        try:
-            sub_nodes = contract['nodes']
-        except:
-            sub_nodes = []
-        state_vars, all_contracts_dict, all_vars = variable_unrolling_new(
-            sub_nodes, all_contracts_dict, all_vars)
-        # parent contract
-        if 'baseContracts' in contract:
-            if contract['baseContracts'] != []:
-                # if contract['name'] == contract_name:
-                    lst = linearized_inherit_tree[contract['name']]
-                    parent_vars = []
-                    for parent_cont_name in lst:
-                        if len(all_contracts_dict) != 0:
-                            b_statevars = all_contracts_dict[parent_cont_name]['vars']
-                            for var in b_statevars:
-                                if type(var) == list:
-                                    for va in var:
-                                        if va not in parent_vars:
-                                            parent_vars.append(va)
-                                else:
-                                    if var not in parent_vars:
-                                        parent_vars.append(var)
-                    state_vars = parent_vars + state_vars
-                    parent = lst
-        all_contracts_dict[contract['name']] = {
-            'vars': state_vars, 'parent': parent, 'type': contract['nodeType']}
-    return all_vars, all_contracts_dict, diamonds
 
 def handle_expression_node(exp, out_nodes, node, compiler_version):
     var = exp.split(' = ')[0]  # get left hand operand
