+++ src/state_extraction/state_extractor.py
@@ -1,10 +1,10 @@
 from logging import raiseExceptions
 import pprint
-from src.key_approx_analysis.key_approx_analyzer import extract_slot_details, generate_final_key_approx_results, get_contract_details, get_contract_details_new, key_approx_analyzer
+from src.key_approx_analysis.key_approx_analyzer import extract_slot_details, generate_final_key_approx_results, key_approx_analyzer
 from src.state_extraction.transactions import get_internal_transactions
 from src.state_extraction.transactions import get_transactions
 from src.state_extraction.slot_calculator import calculate_slots
-from src.ast_parsing.ast_parser import generate_ast
+from src.ast_parsing.ast_parser import generate_ast, get_contract_details, get_contract_details_new
 import collections
 import itertools
 import math
